requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7
pandas>=2.2.3
matplotlib>=3.10.1
numpy>=2.2.5
Flask-WTF>=1.2.1
beautifulsoup4>=4.13.0
lxml>=4.9.3
pytest>=7.4.0
flask-caching>=2.1.0
python-dateutil>=2.8.2